requirements.txt
anyio==3.6.1
bidict==0.22.0
click==8.1.3
dnspython==2.2.1
h11==0.14.0
httptools==0.5.0
idna==3.4
pymongo==4.2.0
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.1
PyYAML==6.0
sniffio==1.3.0
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
